+++ frontend/src/App.tsx
@@ -1,19 +1,19 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { AuthProvider } from "./providers/AuthProvider";
-import Home from "./pages/Home";
-import SignUp from "./pages/SignUp";
-import SignIn from "./pages/SignIn";
-import TeamsSelect from "./pages/TeamsSelect";
-import TeamsShow from "./pages/TeamsShow";
-import UsersShow from "./pages/UsersShow";
-import UsersEdit from "./pages/UsersEdit";
-import TasksShow from "./pages/TasksShow";
-import TasksNew from "./pages/TasksNew";
-import DivisionsNew from "./pages/DivisionsNew";
-import TasksEdit from "./pages/TasksEdit";
-import PrivateRoute from "./components/PrivateRoute";
-import PublicRoute from "./components/PublicRoute";
-import CommonLayout from "./components/CommonLayout";
+import Home from "./components/pages/Home";
+import SignUp from "./components/pages/SignUp";
+import SignIn from "./components/pages/SignIn";
+import TeamsSelect from "./components/pages/TeamsSelect";
+import TeamsShow from "./components/pages/TeamsShow";
+import UsersShow from "./components/pages/UsersShow";
+import UsersEdit from "./components/pages/UsersEdit";
+import TasksShow from "./components/pages/TasksShow";
+import TasksNew from "./components/pages/TasksNew";
+import DivisionsNew from "./components/pages/DivisionsNew";
+import TasksEdit from "./components/pages/TasksEdit";
+import PrivateRoute from "./components/functional/PrivateRoute";
+import PublicRoute from "./components/functional/PublicRoute";
+import CommonLayout from "./components/ui/CommonLayout";
 import { SnackbarProvider } from "./providers/SnackbarProvider";
 import "@fontsource/roboto/300.css";
 import "@fontsource/roboto/400.css";
