+++ frontend/src/components/model/user/UsersEditProfile.tsx
@@ -19,10 +19,10 @@ import {
   Typography,
 } from "@mui/material";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
-import { axiosInstance } from "../utils/axios";
-import { User, UsersUpdateResponse } from "../types";
-import { AuthContext } from "../providers/AuthProvider";
-import { SnackbarContext } from "../providers/SnackbarProvider";
+import { axiosInstance } from "../../../utils/axios";
+import { User, UsersUpdateResponse } from "../../../types";
+import { AuthContext } from "../../../providers/AuthProvider";
+import { SnackbarContext } from "../../../providers/SnackbarProvider";
 
 type Props = {
   user: User;
