+++ frontend/src/components/pages/Home.tsx
@@ -3,10 +3,10 @@ import { useNavigate, Link } from "react-router-dom";
 import Cookies from "js-cookie";
 import { Button, Grid, Typography } from "@mui/material";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
-import { axiosInstance } from "../utils/axios";
-import { AuthResponse } from "../types";
-import { AuthContext } from "../providers/AuthProvider";
-import { SnackbarContext } from "../providers/SnackbarProvider";
+import { axiosInstance } from "../../utils/axios";
+import { AuthResponse } from "../../types";
+import { AuthContext } from "../../providers/AuthProvider";
+import { SnackbarContext } from "../../providers/SnackbarProvider";
 
 const Home: FC = () => {
   const { setIsSignedIn } = useContext(AuthContext);
