+++ frontend/src/components/pages/TasksNew.tsx
@@ -1,28 +1,24 @@
-import { ChangeEvent, FC, useContext, useEffect, useState } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { ChangeEvent, FC, useContext, useState } from "react";
+import { useNavigate } from "react-router-dom";
 import Cookies from "js-cookie";
 import { Button, Grid, TextField, Typography } from "@mui/material";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
-import { axiosInstance } from "../utils/axios";
-import { TasksResponse, EditTask } from "../types";
-import { AuthContext } from "../providers/AuthProvider";
-import { useFetchTask } from "../hooks/useFetchTask";
-import { PriorityTextField } from "../components/PriorityTextField";
-import { DeadlineTextField } from "../components/DeadlineTextField";
-import { SnackbarContext } from "../providers/SnackbarProvider";
+import { axiosInstance } from "../../utils/axios";
+import { TasksResponse, NewTask } from "../../types";
+import { AuthContext } from "../../providers/AuthProvider";
+import { PriorityTextField } from "../model/task/PriorityTextField";
+import { DeadlineTextField } from "../model/task/DeadlineTextField";
+import { SnackbarContext } from "../../providers/SnackbarProvider";
 
-const TasksEdit: FC = () => {
+const TasksNew: FC = () => {
   const { currentUser } = useContext(AuthContext);
   const { handleSetSnackbar } = useContext(SnackbarContext);
   const navigate = useNavigate();
-  const params = useParams<{ id: string }>();
-  const { task: taskData } = useFetchTask();
 
-  const [task, setTask] = useState<EditTask>({
+  const [task, setTask] = useState<NewTask>({
     title: "",
     description: "",
     is_done: false,
-    user_id: 0,
   });
 
   const [deadline, setDeadline] = useState<Date | null>(new Date());
@@ -35,28 +31,35 @@ const TasksEdit: FC = () => {
     setTask({ ...task, [name]: value });
   };
 
-  const handleTasksUpdate = () => {
-    const updateOptions: AxiosRequestConfig = {
-      url: `/tasks/${params.id}`,
-      method: "PATCH",
+  const handleTasksCreate = () => {
+    const options: AxiosRequestConfig = {
+      url: "/tasks",
+      method: "POST",
       headers: {
         "Content-Type": "application/json",
         "access-token": Cookies.get("_access_token") || "",
         client: Cookies.get("_client") || "",
         uid: Cookies.get("_uid") || "",
       },
-      data: task,
+      data: {
+        title: task.title,
+        description: task.description,
+        deadline,
+        priority,
+        is_done: task.is_done,
+        user_id: currentUser?.id,
+      },
     };
 
-    axiosInstance(updateOptions)
+    axiosInstance(options)
       .then((res: AxiosResponse<TasksResponse>) => {
         console.log(res);
 
-        if (res.status === 200) {
+        if (res.status === 201) {
           handleSetSnackbar({
             open: true,
             type: "success",
-            message: "タスクを更新しました",
+            message: "タスクを作成しました",
           });
           navigate(`/users/${currentUser?.id}`, { replace: false });
         }
@@ -72,20 +75,12 @@ const TasksEdit: FC = () => {
       });
   };
 
-  useEffect(() => {
-    if (taskData) {
-      setTask(taskData);
-      setDeadline(taskData.deadline);
-      setPriority(taskData.priority);
-    }
-  }, [taskData]);
-
   return (
     <div>
       <Grid container direction="column" spacing={4}>
         <Grid item>
           <Typography variant="h4" component="div">
-            タスクを編集する
+            タスクを作成する
           </Typography>
         </Grid>
         <Grid item>
@@ -94,7 +89,7 @@ const TasksEdit: FC = () => {
             variant="standard"
             sx={{ width: "30ch" }}
             name="title"
-            value={task?.title}
+            value={task.title}
             onChange={handleInputChange}
           />
         </Grid>
@@ -106,7 +101,7 @@ const TasksEdit: FC = () => {
             rows={4}
             sx={{ width: "55ch" }}
             name="description"
-            value={task?.description}
+            value={task.description}
             onChange={handleInputChange}
           />
         </Grid>
@@ -127,7 +122,7 @@ const TasksEdit: FC = () => {
           />
         </Grid>
         <Grid item>
-          <Button variant="contained" type="submit" onClick={handleTasksUpdate}>
+          <Button variant="contained" type="submit" onClick={handleTasksCreate}>
             完了
           </Button>
         </Grid>
@@ -136,4 +131,4 @@ const TasksEdit: FC = () => {
   );
 };
 
-export default TasksEdit;
+export default TasksNew;
