+++ frontend/src/components/ui/CommonLayout.tsx
