+++ frontend/src/tests/PublicRoute.test.tsx
@@ -1,11 +1,11 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import { rest } from "msw";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
-import PrivateRoute from "../components/PrivateRoute";
-import PublicRoute from "../components/PublicRoute";
+import PrivateRoute from "../components/functional/PrivateRoute";
+import PublicRoute from "../components/functional/PublicRoute";
 import { server } from "../mocks/server";
-import SignIn from "../pages/SignIn";
-import TeamsShow from "../pages/TeamsShow";
+import SignIn from "../components/pages/SignIn";
+import TeamsShow from "../components/pages/TeamsShow";
 import { AuthProvider } from "../providers/AuthProvider";
 
 describe("PublicRoute", () => {
